changed monetary data to double data type

+++ Loja.java
@@ -1,11 +1,11 @@
 public class Loja {
   private String nome;
   private int quantidadeFuncionarios;
-  private int salarioBaseFuncionario;
+  private double salarioBaseFuncionario;
   private Endereco endereco;
   private Data dataFundacao;
 
-  public Loja(String nome, int quantidadeFuncionarios, int salarioBaseFuncionario, Endereco endereco,
+  public Loja(String nome, int quantidadeFuncionarios, double salarioBaseFuncionario, Endereco endereco,
       Data dataFundacao) {
     this.nome = nome;
     this.quantidadeFuncionarios = quantidadeFuncionarios;
@@ -19,7 +19,7 @@ public class Loja {
     this.quantidadeFuncionarios = quantidadeFuncionarios;
     this.endereco = endereco;
     this.dataFundacao = dataFundacao;
-    this.salarioBaseFuncionario = -1; // Valor padrão
+    this.salarioBaseFuncionario = -1.0; // Valor padrão
   }
 
   public String getNome() {
@@ -30,7 +30,7 @@ public class Loja {
     return quantidadeFuncionarios;
   }
 
-  public int getSalarioBaseFuncionario() {
+  public double getSalarioBaseFuncionario() {
     return salarioBaseFuncionario;
   }
 
@@ -50,7 +50,7 @@ public class Loja {
     this.quantidadeFuncionarios = quantidadeFuncionarios;
   }
 
-  public void setSalarioBaseFuncionario(int salarioBaseFuncionario) {
+  public void setSalarioBaseFuncionario(double salarioBaseFuncionario) {
     this.salarioBaseFuncionario = salarioBaseFuncionario;
   }
 
@@ -69,9 +69,9 @@ public class Loja {
         + "\nData de Fundação: " + dataFundacao;
   }
 
-  public int gastosComSalario() {
-    if (this.salarioBaseFuncionario < 0 || this.salarioBaseFuncionario == -1) {
-      return -1; // Indica que o salário base não foi definido
+  public double gastosComSalario() {
+    if (this.salarioBaseFuncionario < 0 || this.salarioBaseFuncionario == -1.0) {
+      return -1.0; // Indica que o salário base não foi definido
     } else {
       return this.quantidadeFuncionarios * this.salarioBaseFuncionario;
     }
